+++ lib/CodeTransform/Tests/Adapter/WorseReflection/Refactor/WorseGenerateMethodTest.php
@@ -4,7 +4,7 @@ namespace Phpactor\CodeTransform\Tests\Adapter\WorseReflection\Refactor;
 
 use Generator;
 use Phpactor\CodeTransform\Tests\Adapter\WorseReflection\WorseTestCase;
-use Phpactor\CodeTransform\Adapter\WorseReflection\Refactor\WorseGenerateMethod;
+use Phpactor\CodeTransform\Adapter\WorseReflection\Refactor\WorseGenerateMember;
 use Phpactor\CodeTransform\Domain\SourceCode;
 use Phpactor\CodeTransform\Domain\Exception\TransformException;
 use Phpactor\CodeBuilder\Adapter\WorseReflection\WorseBuilderFactory;
@@ -13,13 +13,13 @@ use Phpactor\TextDocument\TextDocumentBuilder;
 class WorseGenerateMethodTest extends WorseTestCase
 {
     /**
-     * @dataProvider provideExtractMethod
+     * @dataProvider provideGenerateMember
      */
     public function testGenerateMethod(string $test, ?string $name = null): void
     {
         [$source, $expected, $offset] = $this->sourceExpectedAndOffset(__DIR__ . '/fixtures/' . $test);
 
-        $transformed = $this->generateMethod($source, $offset, $name);
+        $transformed = $this->generateMember($source, $offset, $name);
 
         $this->assertEquals(trim($expected), trim($transformed));
     }
@@ -27,35 +27,36 @@ class WorseGenerateMethodTest extends WorseTestCase
     /**
      * @return Generator<string, list<string>>
      */
-    public function provideExtractMethod(): Generator
+    public function provideGenerateMember(): Generator
     {
-        yield 'string' => [ 'generateMethod1.test' ];
-        yield 'parameter' => [ 'generateMethod2.test' ];
-        yield 'named parameters' => ['generateMethod_namedParams.test'];
-        yield 'typed parameter' => [ 'generateMethod3.test' ];
-        yield 'undeclared parameter' => [ 'generateMethod4.test' ];
-        yield 'expression' => [ 'generateMethod5.test' ];
-        yield 'public accessor in another class' => [ 'generateMethod6.test' ];
-        yield 'public accessor on interface' => [ 'generateMethod7.test' ];
-        yield 'public accessor on interface with namespace' => [ 'generateMethod8.test' ];
+        yield 'string' => [ 'generateMember1.test' ];
+        yield 'parameter' => [ 'generateMember2.test' ];
+        yield 'named parameters' => ['generateMember_namedParams.test'];
+        yield 'typed parameter' => [ 'generateMember3.test' ];
+        yield 'undeclared parameter' => [ 'generateMember4.test' ];
+        yield 'expression' => [ 'generateMember5.test' ];
+        yield 'public accessor in another class' => [ 'generateMember6.test' ];
+        yield 'public accessor on interface' => [ 'generateMember7.test' ];
+        yield 'public accessor on interface with namespace' => [ 'generateMember8.test' ];
+        yield 'imports classes' => [ 'generateMember9.test' ];
+        yield 'static private method' => [ 'generateMember10.test' ];
+        yield 'static public method' => [ 'generateMember11.test' ];
+        yield 'add return type' => [ 'generateMember12.test' ];
+        yield 'add return type with docblock' => [ 'generateMember13.test' ];
+        yield 'add parameter type multiple literals' => [ 'generateMember14.test' ];
+        yield 'nullable parameter inference' => [ 'generateMember15.test' ];
+        yield 'generic parameter inference' => [ 'generateMember16.test' ];
+        yield 'union false' => [ 'generateMember17.test' ];
+        yield 'duplicated type guesses' => [ 'generateMember_duplicateNameGuesses.test' ];
+        yield 'docblock for complex type' => [ 'generateMember_complexTypeDocblock.test' ];
         if (version_compare(PHP_VERSION, '8.1', '>=')) {
-            yield 'public method on enum' => [ 'generateMethod_enum.test', 'play'];
+            yield 'enum' => [ 'generateMember_enumParams.test' ];
+            yield 'backed_enum' => [ 'generateMember_backedEnumParams.test' ];
+            yield 'public method on enum' => [ 'generateMember_enum.test', 'play'];
+            yield 'case on enum' => [ 'generateMember_enumCase.test', 'Foo'];
         }
-        yield 'imports classes' => [ 'generateMethod9.test' ];
-        yield 'static private method' => [ 'generateMethod10.test' ];
-        yield 'static public method' => [ 'generateMethod11.test' ];
-        yield 'add return type' => [ 'generateMethod12.test' ];
-        yield 'add return type with docblock' => [ 'generateMethod13.test' ];
-        yield 'add parameter type multiple literals' => [ 'generateMethod14.test' ];
-        yield 'nullable parameter inference' => [ 'generateMethod15.test' ];
-        yield 'generic parameter inference' => [ 'generateMethod16.test' ];
-        yield 'union false' => [ 'generateMethod17.test' ];
-        yield 'duplicated type guesses' => [ 'generateMethod_duplicateNameGuesses.test' ];
-        yield 'docblock for complex type' => [ 'generateMethod_complexTypeDocblock.test' ];
-        if (version_compare(PHP_VERSION, '8.1', '>=')) {
-            yield 'enum' => [ 'generateMethod_enumParams.test' ];
-            yield 'backed_enum' => [ 'generateMethod_backedEnumParams.test' ];
-        }
+        yield 'private constant on class' => [ 'generateMember_constant.test', 'FOO'];
+        yield 'public constant on class' => [ 'generateMember_constantPublic.test', 'FOO'];
     }
 
     public function testGenerateOnTraitException(): void
@@ -83,21 +84,21 @@ class WorseGenerateMethodTest extends WorseTestCase
             EOT
         ;
 
-        $this->generateMethod($source, 152, 'test_name');
+        $this->generateMember($source, 152, 'test_name');
     }
 
-    private function generateMethod(string $source, int $start, ?string $name): string
+    private function generateMember(string $source, int $start, ?string $name): string
     {
         $worseSourceCode = TextDocumentBuilder::fromPathAndString('file:///source', $source);
         $reflector = $this->reflectorForWorkspace($worseSourceCode);
 
-        $generateMethod = new WorseGenerateMethod(
+        $generateMember = new WorseGenerateMember(
             $reflector,
             new WorseBuilderFactory($reflector),
             $this->updater()
         );
         $sourceCode = SourceCode::fromStringAndPath($source, 'file:///source');
-        $textDocumentEdits = $generateMethod->generateMethod($sourceCode, $start, $name);
+        $textDocumentEdits = $generateMember->generateMember($sourceCode, $start, $name);
 
         $transformed = SourceCode::fromStringAndPath(
             (string) $textDocumentEdits->textEdits()->apply($sourceCode),