+++ lib/WorseReflection/Bridge/TolerantParser/Diagnostics/MissingMemberProvider.php
@@ -16,22 +16,30 @@ use Phpactor\WorseReflection\Core\Diagnostics;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
 use Phpactor\WorseReflection\Core\Inference\Frame;
 use Phpactor\WorseReflection\Core\Inference\NodeContextResolver;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionClassLike;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionEnum;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionMember;
 use Phpactor\WorseReflection\Core\Type\ReflectedClassType;
 
 /**
  * Report if trying to call a class method which does not exist.
  */
-class MissingMethodProvider implements DiagnosticProvider
+class MissingMemberProvider implements DiagnosticProvider
 {
     public function exit(NodeContextResolver $resolver, Frame $frame, Node $node): iterable
     {
-        if ((!$node instanceof CallExpression)) {
+        if (
+            !$node instanceof CallExpression &&
+            !($node instanceof ScopedPropertyAccessExpression && !$node->parent instanceof CallExpression)
+        ) {
             return;
         }
 
         $memberName = null;
-        if ($node->callableExpression instanceof MemberAccessExpression) {
+        $memberType = null;
+        if ($node instanceof ScopedPropertyAccessExpression) {
+            $memberName = $node->memberName;
+        } elseif ($node->callableExpression instanceof MemberAccessExpression) {
             $memberName = $node->callableExpression->memberName;
         } elseif ($node->callableExpression instanceof ScopedPropertyAccessExpression) {
             $memberName = $node->callableExpression->memberName;
@@ -51,26 +59,49 @@ class MissingMethodProvider implements DiagnosticProvider
             return;
         }
 
+        $reflection = $containerType->reflectionOrNull();
+        if (null === $reflection) {
+            return;
+        }
+
+        $memberType = (function (ReflectionClassLike $reflection) use ($node) {
+            if ($reflection instanceof ReflectionEnum) {
+                if ($node instanceof ScopedPropertyAccessExpression) {
+                    return ReflectionMember::TYPE_CASE;
+                }
+                return ReflectionMember::TYPE_METHOD;
+            }
+
+            if ($node instanceof ScopedPropertyAccessExpression) {
+                return ReflectionMember::TYPE_CONSTANT;
+            }
+
+            return ReflectionMember::TYPE_METHOD;
+        })($reflection);
+
         $methodName = $memberName->getText($node->getFileContents());
         if (!is_string($methodName)) {
             return;
         }
         try {
-            $name = $containerType->members()->byMemberType(ReflectionMember::TYPE_METHOD)->get($methodName);
+            $name = $containerType->members()->byMemberType($memberType)->get($methodName);
         } catch (NotFound) {
-            yield new MissingMethodDiagnostic(
+            yield new MissingMemberDiagnostic(
                 ByteOffsetRange::fromInts(
                     $memberName->getStartPosition(),
                     $memberName->getEndPosition()
                 ),
                 sprintf(
-                    'Method "%s" does not exist on class "%s"',
+                    '%s "%s" does not exist on %s "%s"',
+                    ucfirst($memberType),
                     $methodName,
+                    $reflection->classLikeType(),
                     $containerType->__toString()
                 ),
                 DiagnosticSeverity::ERROR(),
                 $containerType->name()->__toString(),
-                $methodName
+                $methodName,
+                $memberType,
             );
         }
     }
@@ -163,6 +194,81 @@ class MissingMethodProvider implements DiagnosticProvider
             assertion: function (Diagnostics $diagnostics): void {
                 Assert::assertCount(1, $diagnostics);
                 Assert::assertEquals('Method "bar" does not exist on class "Foobar"', $diagnostics->at(0)->message());
+            }
+        );
+        if (version_compare(PHP_VERSION, '8.1', '>=')) {
+            yield new DiagnosticExample(
+                title: 'missing enum case',
+                source: <<<'PHP'
+                    <?php
+
+                    enum Foobar
+                    {
+                        case Foo;
+                    }
+
+                    Foobar::Foo;
+                    Foobar::Bar;
+                    PHP,
+                valid: false,
+                assertion: function (Diagnostics $diagnostics): void {
+                    Assert::assertCount(1, $diagnostics);
+                    Assert::assertEquals('Case "Bar" does not exist on enum "Foobar"', $diagnostics->at(0)->message());
+                }
+            );
+            yield new DiagnosticExample(
+                title: 'enum static method call',
+                source: <<<'PHP'
+                    <?php
+
+                    enum Foobar
+                    {
+                    }
+
+                    Foobar::cases();
+                    PHP,
+                valid: false,
+                assertion: function (Diagnostics $diagnostics): void {
+                    Assert::assertCount(0, $diagnostics);
+                }
+            );
+        }
+        yield new DiagnosticExample(
+            title: 'missing constant on class',
+            source: <<<'PHP'
+                <?php
+
+                class Foobar
+                {
+                    const FOO = 'bar';
+                }
+
+                Foobar::FOO;
+                Foobar::BAR;
+                PHP,
+            valid: false,
+            assertion: function (Diagnostics $diagnostics): void {
+                Assert::assertCount(1, $diagnostics);
+                Assert::assertEquals('Constant "BAR" does not exist on class "Foobar"', $diagnostics->at(0)->message());
+            }
+        );
+        yield new DiagnosticExample(
+            title: 'missing property on class is not supported yet',
+            source: <<<'PHP'
+                <?php
+
+                class Foobar
+                {
+                    public int $foo;
+                }
+
+                $f = new Foobar();
+                $f->foo = 12;
+                $f->barfoo = 'string';
+                PHP,
+            valid: false,
+            assertion: function (Diagnostics $diagnostics): void {
+                Assert::assertCount(0, $diagnostics);
             }
         );
     }