+++ lib/WorseReflection/Core/Inference/NodeReflector.php
@@ -10,6 +10,7 @@ use Microsoft\PhpParser\Node\Expression\ObjectCreationExpression;
 use Phpactor\WorseReflection\Bridge\TolerantParser\Reflection\ReflectionAttribute;
 use Phpactor\WorseReflection\Bridge\TolerantParser\Reflection\ReflectionMethodCall;
 use Phpactor\WorseReflection\Bridge\TolerantParser\Reflection\ReflectionObjectCreationExpression as PhpactorReflectionObjectCreationExpression;
+use Phpactor\WorseReflection\Bridge\TolerantParser\Reflection\ReflectionStaticMemberAccess;
 use Phpactor\WorseReflection\Core\Exception\CouldNotResolveNode;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionNode;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionObjectCreationExpression;
@@ -47,16 +48,13 @@ class NodeReflector
         ));
     }
 
-    private function reflectScopedPropertyAccessExpression(Frame $frame, ScopedPropertyAccessExpression $node): ReflectionStaticMethodCall
+    private function reflectScopedPropertyAccessExpression(Frame $frame, ScopedPropertyAccessExpression $node): ReflectionStaticMemberAccess|ReflectionStaticMethodCall
     {
         if ($node->parent instanceof CallExpression) {
             return $this->reflectStaticMethodCall($frame, $node);
         }
 
-        throw new CouldNotResolveNode(sprintf(
-            'Did not know how to reflect node of type "%s"',
-            get_class($node)
-        ));
+        return $this->reflectCaseOrConstant($frame, $node);
     }
 
     private function reflectMemberAccessExpression(Frame $frame, MemberAccessExpression $node): ReflectionMethodCall
@@ -100,6 +98,15 @@ class NodeReflector
     private function reflectAttribute(Frame $frame, Attribute $node): ReflectionNode
     {
         return new ReflectionAttribute(
+            $this->services,
+            $frame,
+            $node
+        );
+    }
+
+    private function reflectCaseOrConstant(Frame $frame, ScopedPropertyAccessExpression $node): ReflectionStaticMemberAccess
+    {
+        return new ReflectionStaticMemberAccess(
             $this->services,
             $frame,
             $node